requirements.txt
asgiref==3.4.1
Django==4.0.1
django-filter==21.1
django-tables2==2.4.1
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
